- import in IntelliJ - bump to java 1.8 in maven and Java 11 in IntelliJ, need to know how to use Java 11 everywhere - ported all terminal IO into Context, maybe this is not the right place, but anyhow is localized in a single class now

+++ tst/org/apache/commons/ognl/test/ASTChainTest.java
@@ -40,10 +40,10 @@ public class ASTChainTest
 
         String expr = "thing[\"x\"].val";
 
-        assertEquals( 1, Ognl.getValue( expr, context, root ) );
+        assertEquals( 1, (int) Ognl.getValue( expr, context, root ) );
 
         Ognl.setValue( expr, context, root, new Integer( 2 ) );
 
-        assertEquals( 2, Ognl.getValue( expr, context, root ) );
+        assertEquals( 2, (int) Ognl.getValue( expr, context, root ) );
     }
 }
